Fix tool parsing for ollama (#1597)
## Description

Ollama requires a specific formatting for JSONschema tools where
parameter types can only be a string, never a list of strings (which is
actually valid JSONschema). To compensate for this we ensure optional
parameters where the type is `['string', 'null']` is transformed to just
`'string'` to be compatible.
This should not have impact, because we provide a list of "required"
parameters with function definitions.

Fixes #1584 

---------

Co-authored-by: Manthan Gupta <manthangupta109@gmail.com>