:bug: fix: checking the class itself instead of it's instance (#1731)
## Description

I have been facing the following error: `TypeError: issubclass() arg 1
must be a class`, each time i make the sub agent generate structured
output based on BaseModel class, apparently the error occurs because
`issubclass()` expects its first argument to be a class, but
`member_agent_run_response.content` is an instance of the
`response_model`, not the class itself.

To fix the error, i passed the class of the object (not the instance) to
`issubclass()`. I used the `type()` function to get the class of the
object as follow:

> issubclass(type(member_agent_run_response.content), BaseModel)