reasoning-update-phi-2507 (#1875)
Changes:

- This update adds DeepSeek as a reasoning model

Next steps:
- Introduce a `reasoning_function` parameter that allows users to pass a
callable function. This function will be executed, and its output will
be added as a message to the Agent. Alternatively, a new parameter can
be introduced to accept a function whose result is directly added as a
message to the Agent. This approach makes the Agent more flexible,
enabling additional use cases beyond reasoning.
- Improve CLI output during `print_response`

---------

Co-authored-by: Ashpreet <ashpreetbedi@users.noreply.github.com>